@@ -32,7 +32,8 @@
     </dependencies>
 
     <properties>
-        <spring-boot.version>2.1.3.RELEASE</spring-boot.version>
+        <spring-boot.version>2.4.1</spring-boot.version>
+        <hikari.version>3.4.5</hikari.version>
     </properties>
 
     <build>
